+++ nostr/src/nips/nip59.rs
@@ -295,11 +295,12 @@ impl GiftWrapBuilder {
 
     /// Add extra tags.
     #[inline]
-    pub fn extra_tags<T>(mut self, tags: T) -> Self
+    pub fn extra_tags<I, T>(mut self, tags: I) -> Self
     where
-        T: IntoIterator<Item = Tag>,
+        I: IntoIterator<Item = T>,
+        T: Into<Tag>,
     {
-        self.extra_tags.extend(tags);
+        self.extra_tags.extend(tags.into_iter().map(Into::into));
         self
     }